Added new notebook (skeletonization_python) to carry out skeletonization using python code. Outputted new data to Individual_cells and skeleton_check
Addition of skeletonization using skimages package:

A new notebook (skeletonization_python.ipynb) was added with code to run skimages.skeletonize on the masks in Individual_cells.

Skeletons were added to the folders in Individual_cells.

A new folder (skeleton_check) was added to compare skeletons and check for branching behaviour